Added speccularity and changed render frequency to 120 fps

+++ Depth3d/Shaders/ShaderProgram.cs
@@ -60,7 +60,7 @@ namespace Depth3d.shaders
         }
 
         // Load uniforms
-        protected void LoadDouble(int location, double value)
+        protected void LoadDouble(int location, float value)
         {
             GL.Uniform1(location, value);
         }
